+++ Microsoft.Azure.Cosmos/src/RequestOptions/RequestOptions.cs
@@ -51,9 +51,11 @@ namespace Microsoft.Azure.Cosmos
         internal virtual ConsistencyLevel? BaseConsistencyLevel { get; set; }
 
         /// <summary>
-        /// This disables all diagnostics for the CosmosDiagnostic in the response.
+        /// This allows user to pass in a custom factory for the diagnostic context.
+        /// A custom implementation can ignore certain calls to avoid additional overhead
+        /// when the information is not required.
         /// </summary>
-        internal CosmosDiagnosticsContext DiagnosticContext { get; set; }
+        internal Func<CosmosDiagnosticsContext> DiagnosticContextFactory { get; set; }
 
         /// <summary>
         /// Fill the CosmosRequestMessage headers with the set properties